+++ CHANGELOG.md
@@ -1,4 +1,11 @@
+## 0.2.0-pre-release
+- Added option to hide columns in mobile version. (> 400px). NOTE: Soon you will be able to choose the screen size you want to hide the columns.
+- Added background color option in line.
+- Added buidler custom to input search.
+- Fix background color filter button.
+
 ## 0.0.1-pre-release
+
 - In this first version, only basic resources were added for building a list with table behavior.
-- Header customization
-- Customization of row cells
+- Header customization.
+- Customization of row cells.